Fix the configurable themes that miss API for custom texts
