+++ models/classes/theme/DefaultTheme.php
@@ -20,7 +20,6 @@
 namespace oat\tao\model\theme;
 
 use oat\tao\helpers\Template;
-use oat\oatbox\Configurable;
 
 /**
  * Class DefaultTheme
@@ -29,10 +28,8 @@ use oat\oatbox\Configurable;
  *
  * @package oat\tao\model\theme
  */
-class DefaultTheme extends Configurable implements Theme
+class DefaultTheme extends ConfigurableTheme
 {
-    private $allTexts;
-
     public function getId()
     {
         return 'default';
@@ -49,86 +46,11 @@ class DefaultTheme extends Configurable implements Theme
 
     /**
      * (non-PHPdoc)
-     * @see \oat\tao\model\theme\Theme::getTemplate()
-     */
-    public function getTemplate($id, $context = Theme::CONTEXT_BACKOFFICE)
-    {
-        switch ($id) {
-    		case 'header-logo' :
-    		    $template = Template::getTemplate('blocks/header-logo.tpl', 'tao');
-    		    break;
-    		case 'footer' :
-    		    $template = Template::getTemplate('blocks/footer.tpl', 'tao');
-    		    break;
-	    	case 'login-message' :
-	    	    $template = Template::getTemplate('blocks/login-message.tpl', 'tao');
-                break;
-	    	default:
-	    	    \common_Logger::w('Unkown template '.$id);
-	    	    $template = null;
-    	}
-    	return $template;
-    }
-
-    /**
-     * (non-PHPdoc)
      * @see \oat\tao\model\theme\Theme::getStylesheet()
      */
     public function getStylesheet($context = Theme::CONTEXT_BACKOFFICE)
     {
         return Template::css('tao-3.css', 'tao');
-    }
-
-    /**
-     * Define all custom text
-     * return [
-     *  'myCustomTextId' => __('My custom text translation');
-     * ];
-     * @return array
-     */
-    protected function initializeTexts()
-    {
-        return [];
-    }
-
-    /**
-     * Allow to set a custom translatable string for a given key
-     * @param String $key
-     * @return string
-     */
-    public function getText($key) {
-        if (empty($this->allTexts)) {
-            $this->allTexts = $this->initializeTexts();
-        }
-        return (array_key_exists($key, $this->allTexts))
-            ? $this->allTexts[$key]
-            : '';
-    }
-
-    /**
-     * Retrieve all custom strings for the given keys
-     * @param String[] $allKeys
-     * @return array
-     */
-    public function getTextFromArray($allKeys) {
-        $allValues = [];
-        if (is_array($allKeys) && ! empty($allKeys)) {
-            forEach ($allKeys as $key) {
-                $allValues[$key] = $this->getText($key);
-            }
-        }
-        return $allValues;
-    }
-
-    /**
-     * Retrieve all existing strings
-     * @return array
-     */
-    public function getAllTexts() {
-        if (empty($this->allTexts)) {
-            $this->allTexts = $this->initializeTexts();
-        }
-        return $this->allTexts;
     }
 
 }
